pyproject.toml
[project]
name = "rag-knowledge-worker"
version = "0.1.0"
description = "RAG Knowledge Worker on Constitution of Estonia"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.4.0",
    "openai>=2.0.0",
    "litellm>=1.81.0",
    "pydantic>=2.0.0",
    "tenacity>=8.0.0",
    "tqdm>=4.65.0",
    "gradio>=6.0.0",
    "python-dotenv>=1.0.0",
    "pymupdf4llm>=0.0.14",
    "pymupdf-layout>=1.0.0",
    "pandas>=2.0.0",
]
